Batch Update
FormsApp projesine Batch Update eklenmesi.

+++ FormsApp/Controllers/HomeController.cs
@@ -140,7 +140,34 @@ public class HomeController : Controller
             return NotFound();
         }
 
+        return View("DeleteConfirm", entity);
+    }
+
+    [HttpPost]
+    public IActionResult Delete(int id, int ProductId)
+    {
+        if(id != ProductId)
+        {
+            return NotFound();
+        }
+
+        var entity = Repository.Products.FirstOrDefault(p => p.ProductId == ProductId);
+        if(entity == null)
+        {
+            return NotFound();
+        }
+
         Repository.DeleteProduct(entity);
+        return RedirectToAction("Index");
+    }
+
+    [HttpPost]
+    public IActionResult EditProducts(List<Product> Products)
+    {
+        foreach (var product in Products)
+        {
+            Repository.EditIsActive(product);
+        }
         return RedirectToAction("Index");
     }
 }
